+++ packages/legend-application-studio-bootstrap/scripts/setup.js
@@ -50,6 +50,9 @@ export const setup = (outputDir) => {
         depot: {
           url: 'http://localhost:6200/depot/api',
         },
+        query: {
+          url: 'http://localhost:9001/query',
+        },
         documentation: {
           url: 'https://legend.finos.org',
           registry: [