+++ packages/legend-application-pure-ide-deployment/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @finos/legend-application-pure-ide-deployment
 
+## 8.33.0
+
+## 8.32.0
+
+## 8.31.0
+
+## 8.30.0
+
 ## 8.29.0
 
 ## 8.28.0