+++ packages/legend-extension-format-morphir/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @finos/legend-extension-format-morphir
 
+## 1.0.37
+
+## 1.0.36
+
+## 1.0.35
+
+## 1.0.34
+
 ## 1.0.33
 
 ## 1.0.32