+++ packages/legend-extension-store-relational/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @finos/legend-extension-store-relational
 
+## 0.0.39
+
+## 0.0.38
+
+## 0.0.37
+
+## 0.0.36
+
 ## 0.0.35
 
 ## 0.0.34