+++ packages/legend-server-sdlc/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @finos/legend-server-sdlc
 
+## 4.1.16
+
 ## 4.1.15
 
 ## 4.1.14