+++ packages/stylelint-config/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @finos/stylelint-config-legend-studio
 
+## 2.0.27
+
 ## 2.0.26
 
 ## 2.0.25